ci to make sure all tests are included, and to make sure there are no lint issues

+++ BulbChipSTM32C071FBPx/run_tests.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+# Auto-update test runners
+python3 manage_tests.py
+if [ $? -ne 0 ]; then
+    echo "Test management script failed. Aborting."
+    exit 1
+fi
+
 # Create output directory
 mkdir -p Tests/build
 rm -f Tests/build/*
